+++ assets/php/actions.php
@@ -39,7 +39,6 @@ if(isset($_POST['newAsset'])){
 	$active 		= isset($_POST['active']) ? 1 : 0;
 
 
-
 	if($name == '') $name = $url;
 
 	if(isset($_POST['multidrop'])){
@@ -77,7 +76,6 @@ if(isset($_POST['newAsset'])){
 		$assetLogName 	= strlen($name) > 35 ? substr($name,0,32)."..." : $name;
 		systemLog('Player', 'Upload asset: '.$assetLogName.' to player '.$playerName, $loginUserID, 1);
 
-
 		if(isset($_POST['multidrop'])){
 			//print_r($images);
 			//print_r();
@@ -85,7 +83,6 @@ if(isset($_POST['newAsset'])){
 			$send	= FALSE;
 			if($set['debug'] == 1) echo 'Send to: '.$playerAddress.'/api/v1/file_asset<br />';
 			if($set['debug'] == 1) print_r($data3);
-
 
 			$url = callURL('POST3', $playerAddress.'/api/v1/file_asset', $data3, $playerID, false);
 			if($set['debug'] == 1) echo 'Response: '.$url.'<br />';
@@ -107,6 +104,12 @@ if(isset($_POST['newAsset'])){
 
 		if($send){
 			$data 										= array();
+			if ($mimetype == 'webpage') {
+				if (strpos($url, "youtube.com") !== false || strpos($url, "youtu.be") !== false) {
+					$mimetype = 'youtube_asset';
+					$data['is_processing'] = 0;
+				}
+			}
 			$data['mimetype'] 				= $mimetype;
 			$data['is_enabled'] 			= $active;
 			$data['name'] 						= $name;